+++ cpp/modules/generators/Billow.h
@@ -2,8 +2,9 @@
 #ifndef BILLOW_H
 #define BILLOW_H
 #include "..\Base.h"
-#include "common\CUDA_Include.h"
+
 namespace noise {
+
 	namespace module {
 
 		// Default parameters
@@ -27,7 +28,7 @@ namespace noise {
 				int octaves = DEFAULT_BILLOW_OCTAVES, float persist = DEFAULT_BILLOW_PERSISTENCE);
 
 			// Get source module count: must be 0, this is a generator and can't have preceding modules.
-			virtual int GetSourceModuleCount() const override;
+			virtual size_t GetSourceModuleCount() const override;
 
 			// Launches the kernel and fills this object's surface object with the relevant data.
 			virtual void Generate() override;
@@ -40,6 +41,7 @@ namespace noise {
 			noiseCfg Attributes;
 
 			noise_t NoiseType;
+
 		};
 	}
 }
