+++ cpp/modules/generators/DecarpientierSwiss.cpp
@@ -4,7 +4,7 @@
 noise::module::DecarpientierSwiss::DecarpientierSwiss(int width, int height, noise_t noise_type, float x, float y, int seed, float freq, float lacun,
 	int octaves, float persist) : Module(width, height), Attributes(seed, freq, lacun, octaves, persist), Origin(x, y), NoiseType(noise_type){}
 
-int noise::module::DecarpientierSwiss::GetSourceModuleCount() const{
+size_t noise::module::DecarpientierSwiss::GetSourceModuleCount() const{
 	return 0;
 }
