+++ cpp/modules/generators/DecarpientierSwiss.h
@@ -26,7 +26,7 @@ namespace noise {
 				int octaves = DEFAULT_DC_SWISS_OCTAVES, float persist = DEFAULT_DC_SWISS_PERSISTENCE);
 
 			// Get source module count: must be 0, this is a generator and can't have preceding modules.
-			virtual int GetSourceModuleCount() const override;
+			virtual size_t GetSourceModuleCount() const override;
 
 			// Launches the kernel and fills this object's surface object with the relevant data.
 			virtual void Generate() override;
