+++ cpp/modules/generators/Voronoi.cpp
@@ -4,9 +4,10 @@
 namespace noise {
 	
 	namespace module {
+
 		Voronoi::Voronoi(int width, int height, voronoi_distance_t cell_dist_func, voronoi_return_t return_t, float freq, float displ, int seed) : Module(width, height), Displacement(displ), Frequency(freq), CellDistanceType(cell_dist_func), ReturnDataType(return_t) {}
 
-		int Voronoi::GetSourceModuleCount() const{
+		size_t Voronoi::GetSourceModuleCount() const{
 			return 0;
 		}
